feat: stop hit proxies on search web UI
Since we look for metadata on sidebar does not make sense to hit the proxies on search, this is the first step to help with other tickets to search in the whole proxy data packages, now is limited to private packages. Unit test are not need it here since is an "assumed" behaviour we don't want anymore. This will clearly help to improve performance on UI.