fix(config): respect the `changePassword` configuration flag
Although the `changePassword` configuration flag was introduced in the
past, it was not taken into account when parsing the config.

With this commit, the value of the `changePassword` flag will be
respected, if it is present in the configuration. Setting the flag to
`true` will turn on the ability to change the password via the web API.
The default value remains `false` to keep backwards compatibility.

> **Note**
> This feature is still experimental and there is no option for changing
> the password in the default web application, but respecting the flag
> allows testing for the respective feature.
> Also, it makes it possible for custom web UIs to take advantage of
> this functionality.